IOS-5773: Increase build time threshold to 250ms and fix build time warnings

+++ iMEGA/Chat Messages/Views/Chat Input Bar/Voice Input/AudioWavesView.swift
@@ -13,7 +13,7 @@ class AudioWavesView: UIView {
         //Width of each view and spacing is 4.0
         let eachViewAndSpacingWidth: CGFloat = 4.0
         let totalWidthAvailable = bounds.width + eachViewAndSpacingWidth
-        let eachBlockWidth = eachViewAndSpacingWidth * 2.0
+        let eachBlockWidth = eachViewAndSpacingWidth * CGFloat(2.0)
         let numberOfViews = Int(round(totalWidthAvailable / eachBlockWidth))
         
         guard stackView.arrangedSubviews.count != numberOfViews else {
